# api/openapi/thinkpixelag.yaml
openapi: 3.1.0
info:
  title: ThinkPixelAG Governance API
  version: 0.1.0-rc.1
  summary: Harness-independent agent governance and run lifecycle API
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
  description: |
    Integration RC contract. All protected operations authenticate a bearer principal or
    workload and authorize the action through policy. Tenant context is derived
    from verified identity and is never accepted from an ordinary request body.
servers:
  - url: https://governance.example.corp
security:
  - bearerAuth: []
...
paths:
...
  /v1/agents:
    get:
      tags: [Agents]
      operationId: listAgents
      summary: List agents visible and invocable by the caller
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Policy-visible active agents with at least one currently approved version, ordered by opaque agent identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentPage'
        '400':
          description: Invalid page size or authenticated cursor
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Caller is not authorized to discover the agent catalog
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/agents/{agent_id}:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    get:
      tags: [Agents]
      operationId: describeAgent
      summary: Describe an agent visible to the caller
      responses:
        '200':
          description: Agent description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          description: Agent is absent, outside the authenticated tenant, inactive, not invocable, or hidden by policy
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/agents/{agent_id}/runs:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    post:
      tags: [Runs]
      operationId: createRun
      summary: Admit a run using a policy-resolved approved agent version
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '201':
          description: Run admitted
          headers:
            Location:
              schema:
                type: string
              description: Canonical run URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '409':
          description: Idempotency conflict, invalid state, or insufficient resources
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/runs/{run_id}:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags: [Runs]
      operationId: getRun
      summary: Get tenant-scoped run status
      responses:
        '200':
          description: Run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          description: Run absent, outside the authenticated tenant, or not visible to the caller
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/runs/{run_id}/signals:
    parameters:
      - $ref: '#/components/parameters/RunId'
    post:
      tags: [Runs]
      operationId: signalRun
      summary: Append a validated signal to a run
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignalRunRequest'
      responses:
        '202':
          description: Signal durably accepted with an ordered event for at-least-once delivery
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunEvent'
        '404':
          description: Run absent, outside the authenticated tenant, or not visible to the caller
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Expected state version is stale or the run is terminal
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/runs/{run_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/RunId'
    post:
      tags: [Runs]
      operationId: cancelRun
      summary: Cancel a run idempotently
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRunRequest'
      responses:
        '200':
          description: Cancelled, replayed, or the terminal transition that won the serialization race
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          description: Run absent, outside the authenticated tenant, or not visible to the caller
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Expected state version is stale or the current nonterminal state cannot be cancelled
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/runs/{run_id}/events:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags: [Runs]
      operationId: streamRunEvents
      summary: Stream ordered resumable run events
      parameters:
        - name: after
          in: query
          description: HMAC-authenticated, run-bound cursor returned as each SSE id; Last-Event-ID may be used on reconnect and must match after when both are supplied
          schema:
            type: string
            maxLength: 512
      responses:
        '200':
          description: Ordered server-sent event stream; id is the resumable cursor, event is the event type, data uses RunEvent JSON, and comment heartbeats keep idle streams live
          headers:
            Cache-Control:
              schema:
                type: string
              description: Always no-cache
          content:
            text/event-stream:
              schema:
                type: string
        '410':
          description: Cursor is outside retention; authoritative refetch required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '400':
          description: Cursor is malformed, unauthenticated, bound to another run, or conflicts with Last-Event-ID
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Run absent, outside the authenticated tenant, or not visible to the caller
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/admin/agents:
    post:
      tags: [Registry administration]
      operationId: createAgent
      summary: Register an agent
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        default:
          $ref: '#/components/responses/Problem'
  /v1/admin/agents/{agent_id}/versions:
    parameters:
      - $ref: '#/components/parameters/AgentId'
    post:
      tags: [Registry administration]
      operationId: registerAgentVersion
      summary: Register an immutable content-addressed agent version
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterAgentVersionRequest'
      responses:
        '201':
          description: Version registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentVersion'
        default:
          $ref: '#/components/responses/Problem'
  /v1/admin/agents/{agent_id}/versions/{version_digest}/approvals:
    parameters:
      - $ref: '#/components/parameters/AgentId'
      - $ref: '#/components/parameters/VersionDigest'
    post:
      tags: [Registry administration]
      operationId: decideAgentVersionApproval
      summary: Approve, reject, deprecate, or revoke an agent version
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
      responses:
        '201':
          description: Approval decision recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Approval'
        default:
          $ref: '#/components/responses/Problem'
  /v1/admin/policies:
    "get":
      "tags":
        - "Policy administration"
      "operationId": "listPolicies"
      "summary": "List policy artifacts"
      "parameters":
        -
          "name": "limit"
          "in": "query"
          "schema":
            "type": "integer"
            "minimum": 1
            "maximum": 100
            "default": 50
        -
          "name": "cursor"
          "in": "query"
          "schema":
            "type": "string"
            "maxLength": 512
      "responses":
        "200":
          "description": "Successful administration operation"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/PolicyList"
        "default":
          "$ref": "#/components/responses/Problem"
    post:
      tags: [Policy administration]
      operationId: uploadPolicy
      summary: Upload a content-addressed signed policy artifact for validation
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyUploadRequest'
      responses:
        '201':
          description: Policy artifact stored or existing digest returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyBundle'
        default:
          $ref: '#/components/responses/Problem'
  /v1/admin/policies/{policy_digest}/activations:
    parameters:
      - $ref: '#/components/parameters/PolicyDigest'
    post:
      tags: [Policy administration]
      operationId: activatePolicy
      summary: Activate an approved policy as a new monotonic activation
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyActivationRequest'
      responses:
        '201':
          description: Activation recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyActivation'
        default:
          $ref: '#/components/responses/Problem'
  /v1/trusted/runs/{run_id}/usage:
    parameters:
      - $ref: '#/components/parameters/RunId'
    post:
      tags: [Resources]
      operationId: recordTrustedUsage
      summary: Record a trusted idempotent metering event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsageEvent'
      responses:
        '202':
          description: Usage accepted or identical event replayed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageReceipt'
        default:
          $ref: '#/components/responses/Problem'
  /v1/trusted/reservations/{reservation_id}/settle:
    parameters:
      - $ref: '#/components/parameters/ReservationId'
    post:
      tags: [Resources]
      operationId: settleReservation
      summary: Settle a child reservation exactly once
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementRequest'
      responses:
        '200':
          description: Established settlement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settlement'
        default:
          $ref: '#/components/responses/Problem'
  /v1/admin/runs/{run_id}/resource-extensions:
    parameters:
      - $ref: '#/components/parameters/RunId'
    post:
      tags: [Resources]
      operationId: extendRunResources
      summary: Add an authorized resource or deadline extension
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceExtensionRequest'
      responses:
        '201':
          description: Extension committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        default:
          $ref: '#/components/responses/Problem'
  /v1/admin/revocations:
    get:
      tags: [Revocations]
      operationId: listRevocations
      summary: List revocations visible to the authorized administrator
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Revocations visible to administrator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevocationPage'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags: [Revocations]
      operationId: createRevocation
      summary: Create a scoped revocation and increment applicable epochs
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRevocationRequest'
      responses:
        '201':
          description: Revocation committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Revocation'
        default:
          $ref: '#/components/responses/Problem'
  /v1/admin/revocations/{revocation_id}/lift:
    parameters:
      - $ref: '#/components/parameters/RevocationId'
    post:
      tags: [Revocations]
      operationId: liftRevocation
      summary: Lift a revocation through a new monotonic change
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LiftRevocationRequest'
      responses:
        '200':
          description: Revocation is lifted and epochs advanced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Revocation'
        default:
          $ref: '#/components/responses/Problem'
  /v1/trusted/revocations/events:
    get:
      tags: [Gateways]
      operationId: streamRevocations
      summary: Stream authenticated ordered revocation changes
      parameters:
        - name: after
          in: query
          description: Authenticated tenant-bound sequence/security-epoch cursor; also accepted through Last-Event-ID.
          schema:
            type: string
            maxLength: 512
        - name: after_sequence
          in: query
          deprecated: true
          description: Initial sequence-only bootstrap cursor; subsequent resumes use the authenticated SSE event ID.
          schema:
            type: integer
            format: int64
            minimum: 0
      responses:
        '200':
          description: SSE stream with bounded writes, heartbeat comments, authenticated event cursors, and a terminal gap event requiring reconciliation.
          content:
            text/event-stream:
              schema:
                type: string
        '410':
          description: Cursor outside retention; reconciliation required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/trusted/revocations/reconcile:
    post:
      tags: [Gateways]
      operationId: reconcileRevocations
      summary: Reconcile gateway sequence and epochs with authoritative state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReconcileRequest'
      responses:
        '200':
          description: Ordered delta or snapshot reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconcileResponse'
        default:
          $ref: '#/components/responses/Problem'
  "/v1/admin/policies/{policy_digest}":
    "get":
      "tags":
        - "Policy administration"
      "operationId": "getPolicy"
      "summary": "Read policy metadata"
      "parameters":
        -
          "name": "policy_digest"
          "in": "path"
          "required": true
          "schema":
            "$ref": "#/components/schemas/Digest"
      "responses":
        "200":
          "description": "Successful administration operation"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/PolicyBundle"
        "default":
          "$ref": "#/components/responses/Problem"
  "/v1/admin/policies/{policy_digest}/source":
    "get":
      "tags":
        - "Policy administration"
      "operationId": "getPolicySource"
      "summary": "Read protected policy source"
      "parameters":
        -
          "name": "policy_digest"
          "in": "path"
          "required": true
          "schema":
            "$ref": "#/components/schemas/Digest"
      "responses":
        "200":
          "description": "Successful administration operation"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/PolicySource"
        "default":
          "$ref": "#/components/responses/Problem"
  "/v1/admin/policy-activations":
    "get":
      "tags":
        - "Policy administration"
      "operationId": "listPolicyActivations"
      "summary": "List activation history"
      "parameters":
        -
          "name": "limit"
          "in": "query"
          "schema":
            "type": "integer"
            "minimum": 1
            "maximum": 100
            "default": 50
        -
          "name": "cursor"
          "in": "query"
          "schema":
            "type": "string"
            "maxLength": 512
      "responses":
        "200":
          "description": "Successful administration operation"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/PolicyActivationList"
        "default":
          "$ref": "#/components/responses/Problem"
  "/v1/admin/policy-activations/current":
    "get":
      "tags":
        - "Policy administration"
      "operationId": "getCurrentPolicyActivation"
      "summary": "Read current activation"
      "responses":
        "200":
          "description": "Successful administration operation"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/PolicyActivation"
        "default":
          "$ref": "#/components/responses/Problem"
  "/v1/admin/policy-drafts":
    "get":
      "tags":
        - "Policy administration"
      "operationId": "listPolicyDrafts"
      "summary": "List latest draft revisions"
      "parameters":
        -
          "name": "limit"
          "in": "query"
          "schema":
            "type": "integer"
            "minimum": 1
            "maximum": 100
            "default": 50
        -
          "name": "cursor"
          "in": "query"
          "schema":
            "type": "string"
            "maxLength": 512
      "responses":
        "200":
          "description": "Successful administration operation"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/PolicyDraftList"
        "default":
          "$ref": "#/components/responses/Problem"
    "post":
      "tags":
        - "Policy administration"
      "operationId": "createPolicyDraft"
      "summary": "Create an inert policy draft"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/PolicyDraftSave"
      "parameters":
        -
          "$ref": "#/components/parameters/IdempotencyKey"
      "responses":
        "201":
          "description": "Successful administration operation"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/PolicyDraftRevision"
        "default":
          "$ref": "#/components/responses/Problem"
...
  "/v1/admin/approvals/{approval_id}":
    "get":
      "tags":
        - "Policy administration"
      "operationId": "getLocalApproval"
      "summary": "Read approval state"
      "parameters":
        -
          "name": "approval_id"
          "in": "path"
          "required": true
          "schema":
            "$ref": "#/components/schemas/Identifier"
      "responses":
        "200":
          "description": "Successful administration operation"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/LocalApprovalView"
        "default":
          "$ref": "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or workload assertion
  parameters:
    AgentId:
      name: agent_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
...
    Cursor:
      name: cursor
      in: query
      description: HMAC-authenticated opaque continuation cursor; it conveys no authorization and is re-evaluated on every page
      schema:
        type: string
        maxLength: 512
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    Problem:
      description: RFC 7807 problem response; hidden resources may map to 404
      headers:
        X-Request-ID:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Digest:
      type: string
      pattern: '^sha256:[a-f0-9]{64}$'
...
    AgentPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/Agent'
        next_cursor:
          type: string
          maxLength: 512
...